First api route setup
