+++ components/topnav.js
@@ -1,8 +1,7 @@
 import Link from 'next/link';
 import { Navbar, Nav } from 'react-bootstrap';
-import 'bootstrap/dist/css/bootstrap.min.css';
 
-import logo from '../static/logo.png';
+// import logo from '../static/logo.png';
 
 const TopNav = () => (
   <div>
@@ -10,13 +9,13 @@ const TopNav = () => (
 
       <Link href="/" passHref>
         <Navbar.Brand href="/">
-          <img
+          {/* <img
             alt=""
             src="{logo}"
             width="30"
             height="30"
             className="d-inline-block align-top"
-          />{' '}
+          />{' '} */}
       Form
     </Navbar.Brand>
       </Link>
