+++ server/package.json
@@ -4,6 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
+    "client": "cd ../ && npm start",
+    "server": "nodemon server.js",
+    "dev": "concurrently --kill-others-on-fail  \"npm run server\" \"npm run client\"",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Alex Smith",
